# market_go_api_gateway/api/docs/swagger.yaml
definitions:
...
  handler.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: integer
    type: object
...
  staff_service.StaffTariff:
    properties:
      amount_for_card:
        type: number
      amount_for_cash:
        type: number
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      name:
        type: string
      typ:
        type: string
      updated_at:
        type: string
    type: object
...
  staff_service.TariffCreateReq:
    properties:
      amount_for_card:
        type: number
      amount_for_cash:
        type: number
      name:
        type: string
      typ:
        type: string
    type: object
  staff_service.TariffCreateResp:
    properties:
      msg:
        type: string
    type: object
  staff_service.TariffDeleteResp:
    properties:
      msg:
        type: string
    type: object
  staff_service.TariffUpdateReq:
    properties:
      amount_for_card:
        type: number
      amount_for_cash:
        type: number
      id:
        type: string
      name:
        type: string
      typ:
        type: string
    type: object
  staff_service.TariffUpdateResp:
    properties:
      msg:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /v1/staff-tariffs:
    get:
      consumes:
      - application/json
      description: Get staff-tariffs
      parameters:
      - default: 10
        description: limit for response
        in: query
        name: limit
        type: integer
      - default: 1
        description: page for response
        in: query
        name: page
        type: integer
      - description: search by type
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/staff_service.StaffTariff'
            type: array
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: List staff-tariffs
      tags:
      - Staff-tariffs
    post:
      consumes:
      - application/json
      description: Create a new staff tariff with the provided details
      parameters:
      - description: data of the staff tariff
        in: body
        name: tariff
        required: true
        schema:
          $ref: '#/definitions/staff_service.TariffCreateReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/staff_service.TariffCreateResp'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create a new staff tariff
      tags:
      - Staff-tariffs
  /v1/staff-tariffs/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a staff-tariff by its unique identifier
      parameters:
      - description: StaffTariff ID to retrieve
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/staff_service.TariffDeleteResp'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete a staff-tariff
      tags:
      - Staff-tariffs
    get:
      consumes:
      - application/json
      description: Retrieve a staff-tariff by its unique identifier
      parameters:
      - description: Staff Tariff ID to retrieve
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/staff_service.StaffTariff'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get a staff-tariff by ID
      tags:
      - Staff-tariffs
    put:
      consumes:
      - application/json
      description: Update an existing staff-tariff with the provided details
      parameters:
      - description: Staff Tariff ID to update
        in: path
        name: id
        required: true
        type: string
      - description: Updated data for the staff-tariff
        in: body
        name: staff-tariff
        required: true
        schema:
          $ref: '#/definitions/staff_service.TariffUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/staff_service.TariffUpdateResp'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update an existing staff-tariff
      tags:
      - Staff-tariffs
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"